Refactor thirty year regulation services to housing company services
