+++ backend/hitas/services/thirty_year_regulation.py
@@ -2,24 +2,19 @@ import datetime
 import logging
 from decimal import Decimal
 from itertools import chain
-from typing import Literal, Optional, TypeAlias, TypedDict
+from typing import Optional, TypeAlias, TypedDict
 
 from dateutil.relativedelta import relativedelta
 from django.db import models
-from django.db.models import ExpressionWrapper, F, OuterRef, Q, Subquery, Sum
-from django.db.models.functions import Coalesce, NullIf, TruncMonth
-from rest_framework.exceptions import ValidationError
-from rest_framework.settings import api_settings
+from django.db.models import F, OuterRef, Q, Subquery
 
-from hitas.exceptions import MissingValues, get_hitas_object_or_404
-from hitas.models import Apartment
-from hitas.models._base import HitasModelDecimalField
+from hitas.exceptions import get_hitas_object_or_404
 from hitas.models.apartment_sale import ApartmentSale
 from hitas.models.external_sales_data import ExternalSalesData, SaleData
 from hitas.models.housing_company import HousingCompany, HousingCompanyState, HousingCompanyWithAnnotations
-from hitas.models.indices import MarketPriceIndex, MarketPriceIndex2005Equal100, SurfaceAreaPriceCeiling
-from hitas.services.apartment import aggregate_catalog_prices_where_no_sales, subquery_first_sale_acquisition_price
-from hitas.utils import business_quarter, hitas_calculation_quarter, max_if_all_not_null, roundup, to_quarter
+from hitas.models.indices import SurfaceAreaPriceCeiling
+from hitas.services.housing_company import get_completed_housing_companies, make_index_adjustment_for_housing_companies
+from hitas.utils import business_quarter, hitas_calculation_quarter, roundup, to_quarter
 
 logger = logging.getLogger()
 
@@ -60,7 +55,14 @@ def perform_thirty_year_regulation(calculation_date: datetime.date) -> Regulatio
     logger.info(f"Checking regulation need for housing companies completed before {regulation_month.isoformat()!r}...")
 
     logger.info("Fetching housing companies...")
-    housing_companies = _get_housing_companies_to_check(regulation_month)
+    housing_companies = get_completed_housing_companies(
+        completion_month=regulation_month,
+        states=[
+            HousingCompanyState.LESS_THAN_30_YEARS,
+            HousingCompanyState.GREATER_THAN_30_YEARS_NOT_FREE,
+            HousingCompanyState.READY_NO_STATISTICS,
+        ],
+    )
     if not housing_companies:
         logger.info("No housing companies to check regulation for.")
         logger.info("Regulation check complete!")
@@ -90,7 +92,7 @@ def perform_thirty_year_regulation(calculation_date: datetime.date) -> Regulatio
     )
 
     logger.info("Making index adjustments...")
-    _index_adjustment(housing_companies, calculation_month)
+    make_index_adjustment_for_housing_companies(housing_companies, calculation_month)
 
     logger.info(f"Fetching surface area price ceiling for {calculation_month.isoformat()!r}...")
     surface_area_price_ceiling = get_hitas_object_or_404(SurfaceAreaPriceCeiling, month=calculation_month)
@@ -129,119 +131,6 @@ def perform_thirty_year_regulation(calculation_date: datetime.date) -> Regulatio
     return results
 
 
-def _get_housing_companies_to_check(regulation_month: datetime.date) -> list[HousingCompanyWithAnnotations]:
-    """Get all housing companies that should be checked for regulation continuation.
-
-    :param regulation_month: Fetch only housing companies that have been
-                             completed on or before this month (YYYY-MM-01).
-    """
-    housing_companies: list[HousingCompanyWithAnnotations] = list(
-        HousingCompany.objects.select_related(
-            "postal_code",
-            "financing_method",
-        )
-        .prefetch_related(
-            "real_estates",
-            "real_estates__buildings",
-            "real_estates__buildings__apartments",
-        )
-        .alias(
-            _first_sale_prices=subquery_first_sale_acquisition_price("real_estates__buildings__apartments__id"),
-            _catalog_prices=aggregate_catalog_prices_where_no_sales("real_estates__buildings__apartments__"),
-            _realized_acquisition_price=ExpressionWrapper(
-                Coalesce(Sum("_first_sale_prices"), 0) + F("_catalog_prices"),
-                output_field=HitasModelDecimalField(),
-            ),
-            _total_surface_area=Sum("real_estates__buildings__apartments__surface_area"),
-        )
-        .annotate(
-            completion_date=max_if_all_not_null(
-                ref="real_estates__buildings__apartments__completion_date",
-                inf=datetime.date.max,
-            ),
-            completion_month=TruncMonth("completion_date"),
-            avg_price_per_square_meter=(
-                F("_realized_acquisition_price")
-                # Prevent zero-division errors
-                / (NullIf(F("_total_surface_area"), 0, output_field=HitasModelDecimalField()))
-            ),
-        )
-        .filter(
-            completion_date__lte=regulation_month,
-            state__in=[
-                HousingCompanyState.LESS_THAN_30_YEARS,
-                HousingCompanyState.GREATER_THAN_30_YEARS_NOT_FREE,
-                # TODO: This should not be taken in 'surface area price ceiling 'calculation!
-                HousingCompanyState.READY_NO_STATISTICS,
-            ],
-        )
-        .all()
-    )
-
-    if not housing_companies:
-        return []
-
-    logger.info(f"{len(housing_companies)} housing companies found!")
-    logger.info("Validating housing company apartment prices and surface areas...")
-    _validate_prices_and_surface_areas(housing_companies)
-
-    return housing_companies
-
-
-def _validate_prices_and_surface_areas(housing_companies: list[HousingCompanyWithAnnotations]) -> None:
-    """
-    Validate that all apartments in the housing company have sales prices or sales catalog prices and surface areas.
-    Raise an error in case they don't.
-    """
-    missing: dict[HousingCompanyNameT, dict[ApartmentAddressT, dict[Literal["price", "surface_area"], bool]]] = {}
-
-    apartments_with_data = Apartment.objects.annotate(
-        missing_prices=ExpressionWrapper(
-            (
-                Q(sales__isnull=True)
-                & (Q(catalog_purchase_price__isnull=True) | Q(catalog_primary_loan_amount__isnull=True))
-            ),
-            output_field=models.BooleanField(),
-        )
-    ).filter(
-        Q(missing_prices=True) | Q(surface_area=None),
-        building__real_estate__housing_company__in=housing_companies,
-    )
-
-    for apartment in apartments_with_data:
-        housing_company_name: HousingCompanyNameT = apartment.housing_company.display_name
-        apartment_address: ApartmentAddressT = apartment.address
-
-        if apartment.missing_prices:  # annotated attribute
-            missing.setdefault(housing_company_name, {})
-            missing[housing_company_name].setdefault(apartment_address, {})
-            missing[housing_company_name][apartment_address]["price"] = True
-
-        if apartment.surface_area is None:
-            missing.setdefault(housing_company_name, {})
-            missing[housing_company_name].setdefault(apartment_address, {})
-            missing[housing_company_name][apartment_address]["surface_area"] = True
-
-    if missing:
-        errors: list[str] = []
-
-        for housing_company_name, apartments_with_data in missing.items():
-            msg = f"Average price per square meter could not be calculated for {housing_company_name!r}:"
-            for apartment_address, flags in apartments_with_data.items():
-                msg += f" Apartment {apartment_address!r} does not have"
-                if flags.get("price"):
-                    msg += " any sales or sales catalog prices"
-                if flags.get("surface_area"):
-                    if flags.get("price"):
-                        msg += " or"
-                    msg += " surface area set"
-                msg += "."
-
-            errors.append(msg)
-
-        raise MissingValues(missing=errors, message="Missing apartment details")
-
-
 def _check_automatic_release(housing_companies: list[HousingCompanyWithAnnotations]) -> list[ComparisonData]:
     """
     Separate out housing companies that are automatically released from regulation.
@@ -265,91 +154,6 @@ def _check_automatic_release(housing_companies: list[HousingCompanyWithAnnotatio
 
     housing_companies[:] = [housing_company for housing_company in housing_companies if housing_company is not None]
     return automatically_released
-
-
-def _index_adjustment(
-    housing_companies: list[HousingCompanyWithAnnotations],
-    calculation_month: datetime.date,
-) -> None:
-    """
-    Make index adjustments for housing company prices.
-    """
-    logger.info("Looking for indices...")
-    indexes = _get_indexes(housing_companies, calculation_month)
-
-    logger.info("All indices found, continuing adjustments...")
-    missing_prices: list[HousingCompanyNameT] = []
-    for housing_company in housing_companies:
-        if housing_company.financing_method.old_hitas_ruleset:
-            calculation_month_index = indexes["old"][calculation_month]
-            completion_month_index = indexes["old"][housing_company.completion_month]
-        else:
-            calculation_month_index = indexes["new"][calculation_month]
-            completion_month_index = indexes["new"][housing_company.completion_month]
-
-        if housing_company.avg_price_per_square_meter in (None, Decimal("0")):
-            missing_prices.append(housing_company.display_name)
-            continue
-
-        housing_company.avg_price_per_square_meter = roundup(
-            housing_company.avg_price_per_square_meter * (calculation_month_index / completion_month_index)
-        )
-
-    if missing_prices:
-        names = ", ".join(f"{housing_company_name!r}" for housing_company_name in missing_prices)
-        raise ValidationError(
-            detail={
-                api_settings.NON_FIELD_ERRORS_KEY: (
-                    f"Average price per square meter zero or missing for these housing companies: {names}. "
-                    f"Index adjustments cannot be made."
-                )
-            },
-        )
-
-
-def _get_indexes(
-    housing_companies: list[HousingCompanyWithAnnotations],
-    calculation_month: datetime.date,
-) -> dict[Literal["old", "new"], dict[datetime.date, Decimal]]:
-    """
-    Fetch indices for index adjustment. Raise an error if not all indices are found.
-    """
-    months_old: set[datetime.date] = set()
-    months_new: set[datetime.date] = set()
-
-    for housing_company in housing_companies:
-        if housing_company.financing_method.old_hitas_ruleset:
-            months_old.add(housing_company.completion_month)
-        else:
-            months_new.add(housing_company.completion_month)
-
-    old_indexes: dict[datetime.date, Decimal] = {}
-    new_indexes: dict[datetime.date, Decimal] = {}
-
-    if months_old:
-        months_old.add(calculation_month)
-        for index in MarketPriceIndex.objects.filter(month__in=months_old).all():
-            old_indexes[index.month] = index.value
-    if months_new:
-        months_new.add(calculation_month)
-        for index in MarketPriceIndex2005Equal100.objects.filter(month__in=months_new).all():
-            new_indexes[index.month] = index.value
-
-    missing_old_indexes = set(months_old).difference(old_indexes) if months_old else set()
-    missing_new_indexes = set(months_new).difference(new_indexes) if months_new else set()
-
-    if missing_old_indexes or missing_new_indexes:
-        errors: list[str] = []
-        if missing_old_indexes:
-            missing_old = ", ".join(f"'{month.strftime('%Y-%m')}'" for month in sorted(missing_old_indexes))
-            errors.append(f"Pre 2011 market price indices missing for months: {missing_old}.")
-        if missing_new_indexes:
-            missing_new = ", ".join(f"'{month.strftime('%Y-%m')}'" for month in sorted(missing_new_indexes))
-            errors.append(f"Post 2011 market price indices missing for months: {missing_new}.")
-
-        raise MissingValues(missing=errors, message="Missing required indices")
-
-    return {"old": old_indexes, "new": new_indexes}
 
 
 def _get_comparison_values(
